feat(v2.0.2): appstate freq u32 -> f32

+++ src/lib.rs
@@ -82,7 +82,7 @@ pub struct Measurements {
 pub enum AppState {
     #[default]
     StandBy,
-    Running(u32), // Frequency in Hz
+    Running(f32), // Frequency in Hz
     Fault,
 }
